fix(console): update captcha backlink and always show captcha toggle
