+++ .changeset/perfect-phones-relax.md
@@ -1,0 +1,9 @@
+---
+"@logto/console": patch
+---
+
+always show enable CAPTCHA toggle
+
+Even if there is no CAPTCHA provider, the toggle will be shown but disabled.
+
+Also the back link of the captcha details page is changed to `/security/captcha`.
