+++ packages/console/src/pages/Security/Captcha/CaptchaForm.module.scss
@@ -1,0 +1,7 @@
+@use '@/scss/underscore' as _;
+
+.description {
+  font: var(--font-body-2);
+  color: var(--color-text-secondary);
+  margin-bottom: _.unit(3);
+}
