+++ packages/console/src/pages/Security/Captcha/index.module.scss
@@ -6,12 +6,6 @@
   flex: 1;
 }
 
-.description {
-  font: var(--font-body-2);
-  color: var(--color-text-secondary);
-  margin-bottom: _.unit(3);
-}
-
 .upsellNotice {
   margin-bottom: _.unit(4);
 }
